Remove duplicate code.
