+++ src/Command/CodeStudio/CodeStudioPipelinesMigrateCommand.php
@@ -7,20 +7,18 @@ use Acquia\Cli\Exception\AcquiaCliException;
 use Acquia\Cli\Output\Checklist;
 use Acquia\DrupalEnvironmentDetector\AcquiaDrupalEnvironmentDetector;
 use AcquiaCloudApi\Endpoints\Account;
-use AcquiaCloudApi\Response\ApplicationResponse;
-use Gitlab\Client;
-use Gitlab\Exception\RuntimeException;
-use Gitlab\HttpClient\Builder;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Filesystem\Path;
 use Symfony\Component\Yaml\Yaml;
-use Webmozart\PathUtil\Path;
 
 /**
  * Class CodeStudioPipelinesMigrateCommand.
  */
 class CodeStudioPipelinesMigrateCommand extends CommandBase {
+
+  use CodeStudioCommandTrait;
 
   protected static $defaultName = 'codestudio:pipelines-migrate';
 
@@ -28,21 +26,6 @@ class CodeStudioPipelinesMigrateCommand extends CommandBase {
    * @var string
    */
   private $appUuid;
-
-  /**
-   * @var \Gitlab\Client
-   */
-  private $gitLabClient;
-
-  /**
-   * @var string
-   */
-  private $gitLabToken;
-
-  /**
-   * @var string
-   */
-  private $gitLabHost;
 
   /**
    * @var string
@@ -75,6 +58,11 @@ class CodeStudioPipelinesMigrateCommand extends CommandBase {
   private $projects;
 
   /**
+   * @var \Acquia\Cli\Output\Checklist
+   */
+  private $checklist;
+
+  /**
    * {inheritdoc}.
    */
   protected function configure() {
@@ -97,53 +85,27 @@ class CodeStudioPipelinesMigrateCommand extends CommandBase {
    */
   protected function execute(InputInterface $input, OutputInterface $output) {
     $this->setMigrateScriptData([]);
-    $this->gitLabHost = $this->getGitLabHost();
-    $this->env = $this->validateEnvironment();
-    $this->gitLabToken = $this->getGitLabToken($this->gitLabHost);
-    $this->getGitLabClient();
-    try {
-      $this->gitLabAccount = $this->gitLabClient->users()->me();
-    }
-    catch (RuntimeException $exception) {
-      $this->io->error([
-        "Unable to authenticate with Code Studio",
-        "Did you set a valid token with the <options=bold>api</> and <options=bold>write_repository</> scopes?",
-        "Try running `glab auth login` to re-authenticate.",
-        "Then try again.",
-      ]);
-      return 1;
-    }
 
-    // Get Cloud access tokens.
-    if (!$input->getOption('key') || !$input->getOption('secret')) {
-      $token_url = 'https://cloud.acquia.com/a/profile/tokens';
-      $this->io->writeln([
-        "",
-        "This will configure AutoDevOps for a Code Studio project using credentials",
-        "(an API Token and SSH Key) belonging to your current Acquia Cloud user account.",
-        "Before continuing, make sure that you're logged into the right Acquia Cloud Platform user account.",
-        "",
-        "<comment>Typically this command should only be run once per application</comment>",
-        "but if your Cloud Platform account is deleted in the future, the Code Studio project will",
-        "need to be re-configured using a different user account.",
-        "",
-        "<options=bold>To begin, visit this URL and create a new API Token for Code Studio to use:</>",
-        "<href=$token_url>$token_url</>",
-      ]);
-    }
-
+    $this->checklist = new Checklist($output);
+    $this->authenticateWithGitLab();
+    $this->writeApiTokenMessage($input);
     $cloud_key = $this->determineApiKey($input, $output);
     $cloud_secret = $this->determineApiSecret($input, $output);
     // We may already be authenticated with Acquia Cloud Platform via a refresh token.
     // But, we specifically need an API Token key-pair of Code Studio.
     // So we reauthenticate to be sure we're using the provided credentials.
     $this->reAuthenticate($cloud_key, $cloud_secret, $this->cloudCredentials->getBaseUri());
-
-    $this->checklist = new Checklist($output);
     $this->appUuid = $this->determineCloudApplication();
+
+    // Get Cloud account.
+    $acquia_cloud_client = $this->cloudApiClientService->getClient();
+    $account_adapter = new Account($acquia_cloud_client);
+    $account = $account_adapter->get();
+    $this->validateRequiredCloudPermissions($acquia_cloud_client, $this->appUuid, $account);
+    $this->setGitLabProjectDescription();
+
     // Get Cloud application.
     $cloud_application = $this->getCloudApplication($this->appUuid);
-    $this->gitLabProjectDescription = "Source repository for Acquia Cloud Platform application <comment>{$this->appUuid}</comment>";
     $project = $this->determineGitLabProject($cloud_application);
 
     // Migrate acquia-pipeline file
@@ -156,17 +118,12 @@ class CodeStudioPipelinesMigrateCommand extends CommandBase {
     $this->migratePostDeploySection();
     $this->io->success([
       "",
-      "*********** Migration completed successfully. *********",
+      "Migration completed successfully.",
       "Created .gitlab-ci.yml and removed acquia-pipeline.yml file.",
       "In order to run Pipeline, push .gitlab-ci.yaml to Main branch of Code Studio project.",
       "Check your pipeline is running in Code Studio for your project."
     ]);
 
-    // Get Cloud account.
-    $acquia_cloud_client = $this->cloudApiClientService->getClient();
-    $account_adapter = new Account($acquia_cloud_client);
-    $account = $account_adapter->get();
-    $this->validateRequiredCloudPermissions($acquia_cloud_client, $this->appUuid, $account);
     return 0;
   }
 
@@ -177,32 +134,6 @@ class CodeStudioPipelinesMigrateCommand extends CommandBase {
    */
   protected function commandRequiresAuthentication(InputInterface $input): bool {
     return FALSE;
-  }
-
-  /**
-   * @throws \Acquia\Cli\Exception\AcquiaCliException
-   */
-  protected function validateEnvironment() {
-    if (!getenv('GITLAB_HOST')) {
-      throw new AcquiaCliException('The GITLAB_HOST environmental variable must be set.');
-    }
-  }
-
-  /**
-   * Search for existing project in code studio.
-   * @param ApplicationResponse $cloud_application
-   *
-   * @return array
-   * @throws \Acquia\Cli\Exception\AcquiaCliException
-   */
-  protected function determineGitLabProject(ApplicationResponse $cloud_application) {
-    $this->projects = $this->gitLabClient->projects()->all(['search' => $cloud_application->uuid]);
-    if (count($this->projects) == 1) {
-      return reset($this->projects);
-    }
-    else {
-      throw new AcquiaCliException("[ERROR] '{$cloud_application->name}' is not configured as Code Studio project.\n [Help] You can configure the '{$cloud_application->name}' project in Code Studio by using 'acli cs:wizard' command \n You can find Code Studio documentation at https://docs.acquia.com/code-studio/getting-started/#create-code-studio-project \n");
-    }
   }
 
   /**
@@ -270,7 +201,7 @@ class CodeStudioPipelinesMigrateCommand extends CommandBase {
    * Migrating varibales to .gitlab-ci.yml file.
    */
   protected function migrateVariablesSection() {
-    $varData = isset($this->acquiaPipelineFileParse['variables'])?$this->acquiaPipelineFileParse['variables']:FALSE;
+    $varData = $this->acquiaPipelineFileParse['variables'] ?? FALSE;
     if ($varData){
       $variables_dump = Yaml::dump(['variables' => $varData]);
       $remove_global = preg_replace('/global:/', '', $variables_dump);
@@ -294,7 +225,7 @@ class CodeStudioPipelinesMigrateCommand extends CommandBase {
    * Migrating build job to .gitlab-ci.yml file.
    */
   protected function migrateBuildSection() {
-    $buildJob=isset($this->acquiaPipelineFileParse['events']['build']['steps'])?$this->acquiaPipelineFileParse['events']['build']['steps']:[];
+    $buildJob= $this->acquiaPipelineFileParse['events']['build']['steps'] ?? [];
     $arrayempty = [];
     if (!empty($buildJob)){
       $response_composer = NULL;
@@ -353,7 +284,7 @@ class CodeStudioPipelinesMigrateCommand extends CommandBase {
    * @param array $arrayempty
    * @param string $keysvar
    */
-  protected function assignBuildStages($arrayempty,$keysvar) {
+  protected function assignBuildStages($arrayempty,$keysvar): array {
     $stages=[
       'setup' => 'Build Drupal',
       'npm run build' => 'Build Drupal',
@@ -402,7 +333,7 @@ class CodeStudioPipelinesMigrateCommand extends CommandBase {
    * Migrating post build job to .gitlab-ci.yml file.
    */
   protected function migratePostDeploySection() {
-    $postDeployJob=isset($this->acquiaPipelineFileParse['events']['post-deploy']['steps'])?$this->acquiaPipelineFileParse['events']['post-deploy']['steps']:[];
+    $postDeployJob= $this->acquiaPipelineFileParse['events']['post-deploy']['steps'] ?? [];
     if(!empty($postDeployJob)){
       $response_launch_ode = NULL;
       foreach ($postDeployJob as $keyss => $valuess){
@@ -445,10 +376,13 @@ class CodeStudioPipelinesMigrateCommand extends CommandBase {
 
   /**
    * Assign post-build job its equivalent stages.
+   *
    * @param array $arrayempty
    * @param string $keysvariable
+   *
+   * @return array
    */
-  protected function assignPostDeployStages($arrayempty,$keysvariable) {
+  protected function assignPostDeployStages(array $arrayempty, string $keysvariable): array {
     $stagess=[
       'launch_ode' => 'Deploy Drupal',
     ];
@@ -470,130 +404,14 @@ class CodeStudioPipelinesMigrateCommand extends CommandBase {
 
   /**
    * Merging the post-deploy job array.
+   *
    * @param array $arrayempty
    */
-  protected function mergePostDeployCode($arrayempty) {
+  protected function mergePostDeployCode(array $arrayempty) {
     $build_dump = Yaml::dump($arrayempty);
     $build_parse = Yaml::parse($build_dump);
     $this->migrateScriptData = array_merge($this->migrateScriptData, $build_parse);
     $this->createGitLabCiFile($this->migrateScriptData);
-  }
-
-  /**
-   * Getting the gitlab token from user.
-   * @param string $gitlab_host
-   *
-   * @return string
-   * @throws \Acquia\Cli\Exception\AcquiaCliException
-   */
-  protected function getGitLabToken(string $gitlab_host): string {
-    if ($this->input->getOption('gitlab-token')) {
-      return $this->input->getOption('gitlab-token');
-    }
-    $process = $this->localMachineHelper->execute([
-      'glab',
-      'config',
-      'get',
-      'token',
-      '--host=' . $gitlab_host,
-    ], NULL, NULL, FALSE);
-    if ($process->isSuccessful() && trim($process->getOutput())) {
-      return trim($process->getOutput());
-    }
-
-    $this->io->writeln([
-      "",
-      "You must first authenticate with Code Studio by creating a personal access token:",
-      "* Visit https://$gitlab_host/-/profile/personal_access_tokens",
-      "* Create a token and grant it both <comment>api</comment> and <comment>write repository</comment> scopes",
-      "* Copy the token to your clipboard",
-      "* Run <comment>glab auth login --hostname=$gitlab_host</comment> and paste the token when prompted",
-      "* Try this command again.",
-    ]);
-
-    throw new AcquiaCliException("Could not determine GitLab token");
-  }
-
-  /**
-   * Getting gitlab host from user.
-   * @return string
-   * @throws \Acquia\Cli\Exception\AcquiaCliException
-   */
-  protected function getGitLabHost(): string {
-    $process = $this->localMachineHelper->execute([
-      'glab',
-      'config',
-      'get',
-      'host',
-    ], NULL, NULL, FALSE);
-    if (!$process->isSuccessful()) {
-      throw new AcquiaCliException("Could not determine GitLab host: {error_message}", ['error_message' => $process->getErrorOutput()]);
-    }
-    $output = trim($process->getOutput());
-    $url_parts = parse_url($output);
-    if (!array_key_exists('scheme', $url_parts) && !array_key_exists('host', $url_parts)) {
-      // $output looks like code.cloudservices.acquia.io.
-      return $output;
-    }
-    // $output looks like http://code.cloudservices.acquia.io/.
-    return $url_parts['host'];
-  }
-
-  /**
-   * @param \AcquiaCloudApi\Connector\Client $acquia_cloud_client
-   * @param string|null $cloud_application_uuid
-   * @param \AcquiaCloudApi\Response\AccountResponse $account
-   *
-   * @throws \Acquia\Cli\Exception\AcquiaCliException
-   */
-  protected function validateRequiredCloudPermissions(\AcquiaCloudApi\Connector\Client $acquia_cloud_client, ?string $cloud_application_uuid, \AcquiaCloudApi\Response\AccountResponse $account): void {
-    $required_permissions = [
-      "deploy to non-prod",
-      # Add SSH key to git repository
-      "add ssh key to git",
-      # Add SSH key to non-production environments
-      "add ssh key to non-prod",
-      # Add a CD environment
-      "add an environment",
-      # Delete a CD environment
-      "delete an environment",
-      # Manage environment variables on a non-production environment
-      "administer environment variables on non-prod",
-    ];
-
-    $permissions = $acquia_cloud_client->request('get', "/applications/{$cloud_application_uuid}/permissions");
-    $keyed_permissions = [];
-    foreach ($permissions as $permission) {
-      $keyed_permissions[$permission->name] = $permission;
-    }
-    foreach ($required_permissions as $name) {
-      if (!array_key_exists($name, $keyed_permissions)) {
-        throw new AcquiaCliException("The Acquia Cloud account {account} does not have the required '{name}' permission. Please add the permissions to this user or use an API Token belonging to a different Acquia Cloud user.", [
-          'account' => $account->mail,
-          'name' => $name
-        ]);
-      }
-    }
-  }
-
-  /**
-   * @return \Gitlab\Client
-   */
-  protected function getGitLabClient(): Client {
-    if (!isset($this->gitLabClient)) {
-      $gitlab_client = new Client(new Builder(new \GuzzleHttp\Client()));
-      $gitlab_client->setUrl('https://' . $this->gitLabHost);
-      $gitlab_client->authenticate($this->gitLabToken, Client::AUTH_OAUTH_TOKEN);
-      $this->setGitLabClient($gitlab_client);
-    }
-    return $this->gitLabClient;
-  }
-
-  /**
-   * @param Client $client
-   */
-  public function setGitLabClient($client) {
-    $this->gitLabClient = $client;
   }
 
   /**
